Don't generate glean_usage tables for deprecated apps
- Deprecated app_variants will no longer have their glean_usage tables
  updated (for the app-id tables/views)
- Deprecated apps will have all app_variants marked as deprecated, and
  will then not be updated (for the app tables/views)
